Fix bug in CTAS when temporary directory is not used
When writing to a new table in Hive connector, it asserts that the target
directory doesn't already exist for every partition (or once in the case of
unpartitioned table), which is doomed to fail when a temporary directory is not
used. This check is now removed when it is detected that temporary directory is
not used.