Add parsing error test to distributed queries
This reproduces the issue fixed in 0498c0df5aff14c9f7f88db1fe2e43ed3470aa5b

None of the existing tests catch it because it only propagates
to the caller if they happen during parsing (which causes the
QueryMonitor to be called inline). Errors that occur during
planning and analysis call the QueryMonitor in a background thread,
so they go to the logs, instead.