Generalize BIGINT to DOUBLE comparison translation
DomainTranslator extracts the TupleDomain for the simple
comparisons (>=, <=, =, <>, etc.). TupleDomain is further
used by connectors to implement the partition pruning
based on the query predicates.

If the column is compared with the value of narrower type
(column_double >= BIGINT '1'), the value can be easily
coerced to the wider type (column_double >= DOUBLE '1.0'),
and no workarounds are needed.

If the column is compared with the value of wider type
(column_integer >= DOUBLE '1.0'), the value must be
rounded to the narrower type.

Rounding algorithm for `column_integer >= const_double`
comparation was already implemented. This commit is a
generalization of that algoritm. Now in order to make some
types pair elligable for rounding and domain translation
the `SATURATED_FLOOR_CAST` operator must be registered for that
types pair.

Resolves: #5013