Add new file based configuration for resource groups
This replaces the dependency on the queueing configuration with a new
config file designed for resource groups. It also allows configuring the
soft limit on memory a group may use. After this limit is reached, new
queries will become queued, even if the "running queries" limit has not
been reached.