Remove static methods from TypeRegistry
Change canCoerce, getCommonSuperType and isTypeOnlyCoercion methods
signature to only accept Type. Make those methods non static and extract
them to the TypeManager interface.

Introduce transitional canCoerce and getCommonSuperType methods
that accept (Type, TypeSignature). Those methods are needed by SignatureBinder
to bind Types to the parametrized signatures. Those methods are going to be removed
in a very next commit.

As isTypeOnlyCoercion is not available by a static reference anymore it can't
be used in TranslationMap directly. Instead of that all the typeOnlyCoercions
must be determined on the Analyze phase, and must be stored into Analysis, so
TranslationMap can take the info about typeOnlyCoercions directly from the
analysis.