Add HDFS impersonation product tests
Add product tests that verify that HDFS impersonation is either enabled or
disabled.

To verify HDFS impersonation a table is created using the Hive connector.
If HDFS impersonation is enabled table data should belong to the Presto JDBC
user, otherwise it should belong to the Hadoop user defined
in Presto configuration.

These tests are profile specific, and can't be run simultaneously on the
product tests environment. In order to exclude such tests from a regular
test suite, that is being run on all the environments, the `profile_specific`
test group has been introduced. This group should be explicitly excluded
for the regular test runs, along with the `quarantine` and `big_query` groups.
Then either the `hdfs_impersonation` or `hdfs_no_impersonation` group should
be included based on the environment configuration we are going to run
product tests on.