Make PartitionedLookupSourceSupplier closeable
ParallelHashBuildOperators will block until the `LookupSourceSupplier` is no longer needed,
and then close it. This states clear ownership over lookupSourceSupplier and prepares for
two future changes:

1. Closing spilling resources

In previous model there had to be two operators (both HashBuilder and LookupJoin) to manage
the closing `LookupSourceSupplier`. Which one should do it depended on query stage.

2. Memory revocation

As above, when there are multiple owners of memory or the task context owns memory, there is
no single entity that can be responsible for revoking the memory reserved for
`LookupSourceSupplier`.